+++ app/scripts/corpora/corpus-set.ts
@@ -88,9 +88,7 @@ export class CorpusSet {
         const attrs = this.map(function (corpus) {
             // Set the is_struct_attr flag for all struct attributes
             Object.values(corpus["struct_attributes"]).forEach((attr) => (attr["is_struct_attr"] = true))
-            // if a position attribute is declared as structural, include here
-            const posAttrs = pickBy(corpus.attributes, (val, key) => val["is_struct_attr"])
-            return { ...posAttrs, ...corpus["struct_attributes"] }
+            return { ...corpus["struct_attributes"] }
         })
         const rest = objectUnion(attrs)